Specifying dimensions to fix custom icons
Sortable grid in this instance relies on icons being exactly 40px in width (we'll just toss in height in there too since W3 loves that) or it breaks. Custom module icons vary in size and break the main layout so we fix that by declaring the dimensions. We also add a bit of padding between the two rows (5px top and bottom in .sortable.grid li ul li) so when the amount of module icons within the grid overflows to a second row they don't hug each other.